requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.12
pillow==10.4.0
numpy==2.0.2
opencv-python==4.10.0.84
requests==2.32.3

torch==2.4.1
transformers==4.44.2

gTTS==2.5.3
SpeechRecognition==3.10.4
